refactor: Update code with removed rules for personal prroperty editor.
